+++ src/main/java/com/example/kotrip/dto/tour/TourInfoDto.java
@@ -11,10 +11,10 @@ public class TourInfoDto {
     private String title;
     private String imageUrl;
     private String addr1;
-    private BigDecimal mapX;
-    private BigDecimal mapY;
+    private Double mapX;
+    private Double mapY;
 
-    public TourInfoDto(int id, String title, String imageUrl, String addr1, BigDecimal mapX, BigDecimal mapY) {
+    public TourInfoDto(int id, String title, String imageUrl, String addr1, Double mapX, Double mapY) {
         this.id = id;
         this.title = title;
         this.imageUrl = imageUrl;
